- The existing (old) commit batching algorithm was stabilised with eliminated random intervals by setting the new server knob COMMIT_BATCH_RANDOMIZE_INTERVAL to false
- Introduced the new commit batching algorithm based on the maximum number of commit batches in progress set with the server knob COMMIT_BATCH_MAX_IN_PROGRESS > 0
- Increased the default commit batching limits:
    - COMMIT_TRANSACTION_BATCH_INTERVAL_MAX from 20 ms to 200 ms
    - COMMIT_TRANSACTION_BATCH_INTERVAL_LATENCY_FRACTION from 0.1 to o.3  (used only with the old algorithm)
    - COMMIT_TRANSACTION_BATCH_BYTES_MIN, COMMIT_TRANSACTION_BATCH_BYTES_MAX, COMMIT_TRANSACTION_BATCH_BYTES_SCALE_BASE from 100 kb to 3 Mb
    - LATENCY_METRICS_LOGGING_INTERVAL from 60 s to 10 s